move DeleteVersionListIterator -> s3DeleteVersionListIterator and into s3-buckets.go
